[chore] route props 전달 방식 변경
