+++ app/(client)/(anon)/spots/[spotId]/tips/components/TipImage.module.scss
@@ -7,5 +7,6 @@
     width: 9.375rem;
     height: 9.375rem;
     border-radius: 1rem;
+    object-fit: cover;
   }
 }
