+++ app/api/admin/spots/[id]/route.ts
@@ -3,7 +3,7 @@ import { GetSpotByIdUseCase } from "@/application/usecases/admin/spot/GetSpotsBy
 import { UpdateSpotUseCase } from "@/application/usecases/admin/spot/UpdateSpotUseCase";
 import { DeleteSpotUseCase } from "@/application/usecases/admin/spot/DeleteSpotUseCase";
 import { PgTicketRepository } from "@/infrastructure/repositories/PgTicketRepository";
-import  SpotRepository  from "@/domain/repositories/SpotRepository";
+import SpotRepository from "@/domain/repositories/SpotRepository";
 import { TicketRepository } from "@/domain/repositories/TicketRepository";
 import PgSpotRepository from "@/infrastructure/repositories/PgSpotRepository";
 
@@ -74,10 +74,13 @@ export async function PATCH(req: Request) {
   }
 }
 
-export async function DELETE(req: Request) {
+export async function DELETE(
+  req: Request,
+  props: { params: Promise<{ id: string }> }
+) {
   try {
-    const url = new URL(req.url);
-    const id = url.pathname.split("/").pop();
+    const params = await props.params;
+    const id = params.id;
 
     if (!id) {
       return NextResponse.json(
